--- lib/widgets/export_private_modal.dart
@@ -1,116 +1,0 @@
-import 'package:citizenwallet/theme/colors.dart';
-import 'package:citizenwallet/widgets/chip.dart';
-import 'package:citizenwallet/widgets/dismissible_modal_popup.dart';
-import 'package:citizenwallet/widgets/header.dart';
-import 'package:flutter/cupertino.dart';
-import 'package:go_router/go_router.dart';
-
-class ExportPrivateModal extends StatelessWidget {
-  final String title;
-  final String copyLabel;
-  final void Function() onCopy;
-
-  const ExportPrivateModal({
-    Key? key,
-    this.title = 'Wallet',
-    required this.copyLabel,
-    required this.onCopy,
-  }) : super(key: key);
-
-  void handleDismiss(BuildContext context) {
-    GoRouter.of(context).pop();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final height = MediaQuery.of(context).size.height;
-
-    return DismissibleModalPopup(
-      modaleKey: 'qr-modal',
-      maxHeight: height,
-      paddingSides: 10,
-      onUpdate: (details) {
-        if (details.direction == DismissDirection.down &&
-            FocusManager.instance.primaryFocus?.hasFocus == true) {
-          FocusManager.instance.primaryFocus?.unfocus();
-        }
-      },
-      onDismissed: (_) => handleDismiss(context),
-      child: GestureDetector(
-        onTap: () => FocusManager.instance.primaryFocus?.unfocus(),
-        child: CupertinoPageScaffold(
-          backgroundColor: ThemeColors.uiBackground.resolveFrom(context),
-          child: SafeArea(
-            child: Flex(
-              direction: Axis.vertical,
-              children: [
-                Header(
-                  title: title,
-                  actionButton: CupertinoButton(
-                    padding: const EdgeInsets.all(5),
-                    onPressed: () => handleDismiss(context),
-                    child: Icon(
-                      CupertinoIcons.xmark,
-                      color: ThemeColors.touchable.resolveFrom(context),
-                    ),
-                  ),
-                ),
-                Expanded(
-                  child: ListView(
-                    scrollDirection: Axis.vertical,
-                    children: [
-                      const SizedBox(
-                        height: 10,
-                      ),
-                      const Text(
-                        'Private Key',
-                        textAlign: TextAlign.center,
-                        style: TextStyle(
-                            fontSize: 24, fontWeight: FontWeight.bold),
-                      ),
-                      const SizedBox(
-                        height: 10,
-                      ),
-                      const Text(
-                        'Keep this key safe, anyone with access to it has access to the account.',
-                        textAlign: TextAlign.center,
-                        style: TextStyle(
-                            fontSize: 16, fontWeight: FontWeight.normal),
-                      ),
-                      const SizedBox(
-                        height: 20,
-                      ),
-                      Row(
-                        mainAxisAlignment: MainAxisAlignment.center,
-                        children: [
-                          Chip(
-                            onTap: onCopy,
-                            copyLabel,
-                            color:
-                                ThemeColors.subtleEmphasis.resolveFrom(context),
-                            textColor:
-                                ThemeColors.touchable.resolveFrom(context),
-                            suffix: Icon(
-                              CupertinoIcons.square_on_square,
-                              size: 14,
-                              color: ThemeColors.touchable.resolveFrom(context),
-                            ),
-                            borderRadius: 15,
-                            maxWidth: 150,
-                          ),
-                        ],
-                      ),
-                      const SizedBox(
-                        height: 20,
-                      ),
-                    ],
-                  ),
-                ),
-              ],
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-}
